% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/read_bed_file.R
\name{read_bed_file}
\alias{read_bed_file}
\title{Read BED File}
\usage{
read_bed_file(bed_file_path)

read_bed_file(bed_file_path)
}
\arguments{
\item{bed_file_path}{A string giving the file path of the BED file to read.}
}
\value{
A tibble with the columns V3, V4, X5, and amplikon from the BED file.

A tibble with the columns V3, V4, X5, and amplikon from the BED file.
}
\description{
Reads a BED file and returns a tibble with columns V3, V4, X5, and amplikon.

Reads a BED file and returns a tibble with columns V3, V4, X5, and amplikon.
}
\examples{
bed_data <- read_bed_file("path/to/your/bed_file.bed")
bed_data <- read_bed_file("path/to/your/bed_file.bed")
}
